+++ lib/apache_age/edge.rb
@@ -1,0 +1,5 @@
+module ApacheAge
+  class Edge
+    include ApacheAge::Entities::Edge
+  end
+end