+++ spec/dummy/app/nodes/character.rb
@@ -1,0 +1,13 @@
+class Character
+  include ApacheAge::Entities::Node
+
+  attribute :name, :string
+
+  validates :name, presence: true
+
+  # custom unique node validator (remove any attributes that are NOT important to uniqueness)
+  validates_with(
+    ApacheAge::Validators::UniqueNode,
+    attributes: [:name]
+  )
+end